requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.2.0
